TASK_1/Experimentation/requirements.txt
python
torch
torchvision
transformers
einops
scikit-learn
numpy
Pillow
scipy
torchattacks
tqdm
random